+++ CMakeLists.txt
@@ -34,7 +34,7 @@ add_executable(tests
     src/tests.cpp
     src/bitfield/bitfield_tests.cpp
     src/manager_tests.cpp
-    src/pubsub/eventmanager_tests.cpp
+    src/events/EventManager_tests.cpp
 )
 add_dependencies(tests doctest)
