--- src/pubsub/PubSub.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "EventManager.hpp"
-#include "ECSEvents.hpp"
